apps/sticky/pyproject.toml
[project]
name = "sticky"
version = "0.1.0"
description = "Telegram sticker analytics + dynamic-pack manager — reads Telegram-macOS locally, builds native sticker packs"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Sticky" }]
keywords = ["telegram", "stickers", "analytics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "httpx>=0.27",
    "keyring>=25.0",
    "sqlcipher3-wheels>=0.5.7",
    "pycryptodomex>=3.20",
    "mmh3>=4.1,<5",
    "pydantic>=2.7",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "greenlet>=3.0",
    "jinja2>=3.1",
    "pyjwt>=2.9",
]

[project.scripts]
sticky = "sticky.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sticky"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
